+++ src/services/criarAnuncio.js
@@ -3,7 +3,7 @@ import { ref, uploadBytes, getDownloadURL } from "firebase/storage";
 import { collection, doc, setDoc } from "firebase/firestore";
 
 export default async function addDados(dados) {
-     const { email, nomeUsuario, foto, nome, descricao, localizacao, raca, sexo } = dados;
+     const { email, nomeUsuario, foto, nome, idade, descricao, localizacao, raca, sexo } = dados;
      console.log("Dados do usuário: " + email + "-" + nomeUsuario);
 
      try {
@@ -29,6 +29,7 @@ export default async function addDados(dados) {
                email,
                nomeUsuario,
                nome, 
+               idade,
                descricao, 
                localizacao,
                raca,
